Remove unnecessary casts.

+++ src/main/java/pam/pamhc2crops/items/ItemPamGrain.java
@@ -33,7 +33,7 @@ public class ItemPamGrain extends BlockNamedItem {
 
 		ItemStack stack = player.getHeldItem(hand);
 
-		if (!entity.world.isRemote && !entity.isChild() && entity instanceof AgeableEntity && (int) ((AgeableEntity) entity).getGrowingAge() == 0) {
+		if (!entity.world.isRemote && !entity.isChild() && entity instanceof AgeableEntity && ((AgeableEntity) entity).getGrowingAge() == 0) {
 			if (entity instanceof CowEntity) {
 				if (((CowEntity) entity).isInLove()) {
 					return ActionResultType.FAIL;
@@ -60,7 +60,7 @@ public class ItemPamGrain extends BlockNamedItem {
 		if (entity.isChild()) {
 			if (!player.isCreative())
 				stack.shrink(1);
-			((AgeableEntity) entity).ageUp((int) ((float) (-((AgeableEntity) entity).getGrowingAge() / 20) * 0.1F),
+			((AgeableEntity) entity).ageUp((int) (-((AgeableEntity) entity).getGrowingAge() / 20 * 0.1F),
 					true);
 			return ActionResultType.PASS;
 		}
